deleted action creators folder and consolidated into reducder files
