--- app/action-creators/constants.jsx
@@ -1,3 +1,0 @@
-// User
-export const INITIALIZE_USERS = 'INITIALIZE_USERS'
-export const CREATE_USER = 'CREATE_USER'
